mobilenet-ssd-tflite-object-detection/requirements.txt
numpy==2.2.2
opencv_python==4.11.0.86
Pillow==11.1.0
tensorflow==2.18.0
